Avoid platform-related tzinfo-data bundler warnings
* We can't just use the `bundle config --local disable_platform_warnings
true` command [1] inside the nix-shell, because the bundler
configuration directory is inside the nix store.

* Bundix has a mechanism [2] for including credentials for gem sources
from a supplied bundler configuration file. We can't just use this as it
stands, because it only looks for credentials and not other bundler
configuration settings. However, we could look at enhancing this
mechanism to support the `disable_platform_warnings` setting. But this
seems like a lot of work for a very minor issue!

* Setting `BUNDLE_DISABLE_PLATFORM_WARNINGS` [3] in the user-level
bundler configuration seems like the simplest option. This is
reasonable, because we know the platform is Linux for all Rails projects
on the VM.

[1]: https://maxwellholder.com/2019/01/04/tzinfo-data-warning.html
[2]: https://github.com/nix-community/bundix/blob/5313a047f318a2a44f0377ee3d53067bbc05de36/lib/bundix/source.rb#L36
[3]: https://bundler.io/v2.1/man/bundle-config.1.html#LIST-OF-AVAILABLE-KEYS